Fix game-history again
