+++ frontend/smart-kickers-game/src/StartGame.js
@@ -1,10 +1,19 @@
 import React from 'react';
 import './StartGame.css';
-import useCurrentGameplay from './hooks/useCurrentGameplay';
 import { Button } from './components/Button/Button';
+import { useNavigate } from 'react-router-dom';
+import { useGameDataContext } from './contexts/GameDataContext';
 
 export default function StartGame() {
-  const { handleStartGame } = useCurrentGameplay();
+  const navigate = useNavigate();
+  const { start, handleResetGame, resetGoalsArray, setIsGameStarted } = useGameDataContext();
+  const handleStartGame = () => {
+    setIsGameStarted(true);
+    handleResetGame();
+    resetGoalsArray();
+    start();
+    navigate('/game');
+  };
   return (
     <>
       <h1>Smart Kickers</h1>
