+++ frontend/smart-kickers-game/src/hooks/useCurrentGameplay.js
@@ -1,17 +1,11 @@
 import { useEffect, useRef, useState } from 'react';
 import { Goal, TeamID } from '../constants/score.js';
-import { useStopwatch } from 'react-timer-hook';
 import { useGameDataContext } from '../contexts/GameDataContext';
-import { resetGame } from '../apis/resetGame';
-import { useNavigate } from 'react-router-dom';
 
 export default function useCurrentGameplay() {
-  const navigate = useNavigate();
-  const [isVisible, setIsVisible] = useState(false);
+  const { blueScore, whiteScore, setFinalScores, goalsArray, minutes, seconds, reset, start, handleResetGame, handleStartGame, setIsGameEnded } =
+    useGameDataContext();
 
-  const { blueScore, whiteScore, setFinalScores, goalsArray, setGoalsArray, setIsGameEnded, setIsGameStarted } = useGameDataContext();
-
-  const { seconds, minutes, start, pause, reset } = useStopwatch({ autoStart: false });
   const ScorePrevious = (value) => {
     const ref = useRef();
     useEffect(() => {
@@ -23,48 +17,31 @@ export default function useCurrentGameplay() {
   const prevWhiteScore = ScorePrevious(whiteScore);
 
   useEffect(() => {
-    if (prevBlueScore > blueScore) {
-      const indexOfLastBlue = goalsArray.indexOf(goalsArray.findLast((e) => e.teamID === TeamID.Team_blue));
-      goalsArray.splice(indexOfLastBlue, 1);
-    } else {
-      goalsArray.push(new Goal(TeamID.Team_blue, 'time: ' + minutes.toString().padStart(2, '0') + ':' + seconds.toString().padStart(2, '0')));
+    if (prevBlueScore !== undefined) {
+      if (prevBlueScore > blueScore) {
+        const indexOfLastBlue = goalsArray.indexOf(goalsArray.findLast((e) => e.teamID === TeamID.Team_blue));
+        goalsArray.splice(indexOfLastBlue, 1);
+      } else {
+        goalsArray.push(new Goal(TeamID.Team_blue, 'time: ' + minutes.toString().padStart(2, '0') + ':' + seconds.toString().padStart(2, '0')));
+      }
     }
   }, [blueScore]);
+
   useEffect(() => {
-    if (prevWhiteScore > whiteScore) {
-      const indexOfLastWhite = goalsArray.indexOf(goalsArray.findLast((e) => e.teamID === TeamID.Team_white));
-      goalsArray.splice(indexOfLastWhite, 1);
-    } else {
-      goalsArray.push(new Goal(TeamID.Team_white, 'time: ' + minutes.toString().padStart(2, '0') + ':' + seconds.toString().padStart(2, '0')));
+    if (prevWhiteScore !== undefined) {
+      if (prevWhiteScore > whiteScore) {
+        const indexOfLastWhite = goalsArray.indexOf(goalsArray.findLast((e) => e.teamID === TeamID.Team_white));
+        goalsArray.splice(indexOfLastWhite, 1);
+      } else {
+        goalsArray.push(new Goal(TeamID.Team_white, 'time: ' + minutes.toString().padStart(2, '0') + ':' + seconds.toString().padStart(2, '0')));
+      }
     }
   }, [whiteScore]);
 
-  const handleStartGame = () => {
-    navigate('/game');
-    resetGoalsArray();
-    handleResetGame();
-    setIsVisible(true);
-    start();
-    setIsGameStarted(true);
-  };
-
-  const resetGoalsArray = () => {
-    setGoalsArray([]);
-  };
-
-  const handleResetGame = () => {
-    resetGame().then((data) => {
-      if (data.error) alert(data.error);
-    });
-    reset();
-    start();
-    resetGoalsArray();
-  };
   const handleEndGame = () => {
     setFinalScores({ blueScore: blueScore, whiteScore: whiteScore });
-    pause();
     setIsGameEnded(true);
   };
 
-  return { isVisible, handleStartGame, handleResetGame, handleEndGame };
+  return { handleStartGame, handleResetGame, handleEndGame, start, reset };
 }
